feat: add chech token exist and save token
