+++ src/models/Token.ts
@@ -5,7 +5,7 @@ export interface IToken {
   symbol: string;
   name: string;
   decimals: string;
-  logo: string;
+  logo?: string;
 }
 
 const Token = new Schema<IToken>({
@@ -13,7 +13,7 @@ const Token = new Schema<IToken>({
   symbol: { type: String, required: true, unique: true },
   name: { type: String, required: true, unique: true },
   decimals: { type: String, required: true, unique: false },
-  logo: { type: String, required: true, unique: true },
+  logo: { type: String, required: false, unique: true },
 });
 
 export default model<IToken>('Token', Token);
